refactor(shared/components): styles 관련 파일명 변경
