+++ packages/shared/components/src/Input/input.tsx
@@ -1,6 +1,6 @@
 import { type HTMLFavolinkProps, favolink, forwardRef } from '@favolink/system';
 import { classNames } from '@favolink/utils';
-import * as styles from './input-styles.css';
+import * as styles from './input.styles.css';
 
 export type InputProps = HTMLFavolinkProps<'input'> & {
   variant?: styles.InputVariant;
